refactor: split CustomerQuotaScreen into components
